Fix various errors and most warnings
--HG--
branch : multiweight

+++ src/Core/AnalysisHandler.cc
@@ -480,20 +480,21 @@ namespace Rivet {
     _eventCounter.get()->setActiveWeightIdx(_defaultWeightIdx);
     double nomwgt = sumOfWeights();
 
-    // the cross section of each weight variation is the nominal cross section
+    // The cross section of each weight variation is the nominal cross section
     // times the sumOfWeights(variation) / sumOfWeights(nominal).
-    // this way the cross section will work correctly
-    for (int iW = 0; iW < numWeights(); iW++) {
-        _eventCounter.get()->setActiveWeightIdx(iW);
-        double s = sumOfWeights() / nomwgt;
-        _xs.get()->setActiveWeightIdx(iW);
-        _xs->addPoint(xs*s, xserr*s);
+    // This way the cross section will work correctly
+    for (size_t iW = 0; iW < numWeights(); iW++) {
+      _eventCounter.get()->setActiveWeightIdx(iW);
+      double s = sumOfWeights() / nomwgt;
+      _xs.get()->setActiveWeightIdx(iW);
+      _xs->addPoint(xs*s, xserr*s);
     }
 
     _eventCounter.get()->unsetActiveWeight();
     _xs.get()->unsetActiveWeight();
     return *this;
   }
+
 
   AnalysisHandler& AnalysisHandler::addAnalysis(Analysis* analysis) {
     analysis->_analysishandler = this;
